pyproject.toml
[tool.poetry]
name = "bavourai"
version = "0.1.0"
description = ""
authors = ["Baron Chibuike <baronchibuike@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
langchain = "^0.3.13"
openai = "^1.58.1"
requests = "^2.32.3"
chromadb = "^0.5.23"
numpy = "^1.26.0"
python-dotenv = "^1.0.1"
langchain-community = "^0.3.13"
pypdf = "^5.1.0"
beautifulsoup4 = "^4.12.3"
motor = "^3.6.0"
loguru = "^0.7.3"
pinecone-client = "^5.0.1"
langchain-ollama = "^0.2.2"
langchain-chroma = "^0.1.4"
sentence-transformers = "^3.3.1"
mem0ai = "^0.1.38"
spacy = "^3.8.3"
transformers = "4.46.1"
pandas = "^2.2.3"
pymongo = "^4.11"
duckdb = "^1.2.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
